% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oSNP2xGenes.r
\name{oSNP2xGenes}
\alias{oSNP2xGenes}
\title{Function to prioritise genes given a list of seed SNPs together with the significance level (e.g. GWAS reported p-values)}
\usage{
oSNP2xGenes(
data,
include.LD = NA,
LD.customised = NULL,
LD.r2 = 0.8,
significance.threshold = 5e-05,
score.cap = 10,
distance.max = 2000,
decay.kernel = c("slow", "constant", "linear", "rapid"),
decay.exponent = 2,
GR.SNP = c("dbSNP_GWAS", "dbSNP_Common", "dbSNP_Single"),
GR.Gene = c("UCSC_knownGene", "UCSC_knownCanonical"),
include.TAD = c("none", "GM12878", "IMR90", "MSC", "TRO", "H1", "MES",
"NPC"),
include.QTL = NA,
QTL.customised = NULL,
include.RGB = NA,
cdf.function = c("empirical", "exponential"),
relative.importance = c(1/3, 1/3, 1/3),
scoring.scheme = c("max", "sum", "sequential"),
verbose = TRUE,
placeholder = NULL,
guid = NULL
)
}
\arguments{
\item{data}{a named input vector containing the sinificance level for
nodes (dbSNP). For this named vector, the element names are dbSNP ID
(or in the format such as 'chr16:28525386'), the element values for the
significance level (measured as p-value or fdr). Alternatively, it can
be a matrix or data frame with two columns: 1st column for dbSNP, 2nd
column for the significance level}

\item{include.LD}{additional SNPs in LD with Lead SNPs are also
included. By default, it is 'NA' to disable this option. Otherwise, LD
SNPs will be included based on one or more of 26 populations and 5
super populations from 1000 Genomics Project data (phase 3). The
population can be one of 5 super populations ("AFR", "AMR", "EAS",
"EUR", "SAS"), or one of 26 populations ("ACB", "ASW", "BEB", "CDX",
"CEU", "CHB", "CHS", "CLM", "ESN", "FIN", "GBR", "GIH", "GWD", "IBS",
"ITU", "JPT", "KHV", "LWK", "MSL", "MXL", "PEL", "PJL", "PUR", "STU",
"TSI", "YRI"). Explanations for population code can be found at
\url{http://www.1000genomes.org/faq/which-populations-are-part-your-study}}

\item{LD.customised}{a user-input matrix or data frame with 3 columns:
1st column for Lead SNPs, 2nd column for LD SNPs, and 3rd for LD r2
value. It is designed to allow the user analysing their pre-calculated
LD info. This customisation (if provided) has the high priority over
built-in LD SNPs}

\item{LD.r2}{the LD r2 value. By default, it is 0.8, meaning that SNPs
in LD (r2>=0.8) with input SNPs will be considered as LD SNPs. It can
be any value from 0.8 to 1}

\item{significance.threshold}{the given significance threshold. By
default, it is set to NULL, meaning there is no constraint on the
significance level when transforming the significance level of SNPs
into scores. If given, those SNPs below this are considered significant
and thus scored positively. Instead, those above this are considered
insigificant and thus receive no score}

\item{score.cap}{the maximum score being capped. By default, it is set
to 10. If NULL, no capping is applied}

\item{distance.max}{the maximum distance between genes and SNPs. Only
those genes no far way from this distance will be considered as seed
genes. This parameter will influence the distance-component weights
calculated for nearby SNPs per gene}

\item{decay.kernel}{a character specifying a decay kernel function. It
can be one of 'slow' for slow decay, 'linear' for linear decay, and
'rapid' for rapid decay. If no distance weight is used, please select
'constant'}

\item{decay.exponent}{an integer specifying a decay exponent. By
default, it sets to 2}

\item{GR.SNP}{the genomic regions of SNPs. By default, it is
'dbSNP_GWAS', that is, SNPs from dbSNP (version 146) restricted to GWAS
SNPs and their LD SNPs (hg19). It can be 'dbSNP_Common', that is,
Common SNPs from dbSNP (version 146) plus GWAS SNPs and their LD SNPs
(hg19). Alternatively, the user can specify the customised input. To do
so, first save your RData file (containing an GR object) into your
local computer, and make sure the GR object content names refer to
dbSNP IDs. Then, tell "GR.SNP" with your RData file name (with or
without extension), plus specify your file RData path in "placeholder".
Note: you can also load your customised GR object directly}

\item{GR.Gene}{the genomic regions of genes. By default, it is
'UCSC_knownGene', that is, UCSC known genes (together with genomic
locations) based on human genome assembly hg19. It can be
'UCSC_knownCanonical', that is, UCSC known canonical genes (together
with genomic locations) based on human genome assembly hg19.
Alternatively, the user can specify the customised input. To do so,
first save your RData file (containing an GR object) into your local
computer, and make sure the GR object content names refer to Gene
Symbols. Then, tell "GR.Gene" with your RData file name (with or
without extension), plus specify your file RData path in "placeholder".
Note: you can also load your customised GR object directly}

\item{include.TAD}{TAD boundary regions are also included. By default,
it is 'none' to disable this option. Otherwise, inclusion of a TAD
dataset to pre-filter SNP-nGene pairs (i.e. only those within a TAD
region will be kept). TAD datasets can be one of "GM12878"
(lymphoblast), "IMR90" (fibroblast), "MSC" (mesenchymal stem cell)
,"TRO" (trophoblasts-like cell), "H1" (embryonic stem cell), "MES"
(mesendoderm) and "NPC" (neural progenitor cell). Explanations can be
found at \doi{10.1016/j.celrep.2016.10.061}}

\item{include.QTL}{the eQTL supported currently. By default, it is 'NA'
to disable this option. Pre-built eQTL datasets are detailed in
\code{\link{oDefineQTL}}}

\item{QTL.customised}{a user-input matrix or data frame with 4 columns:
1st column for SNPs/eQTLs, 2nd column for Genes, 3rd for eQTL mapping
significance level (p-values or FDR), and 4th for contexts (required
even though only one context is input). Alternatively, it can be a file
containing these 4 columns. It is designed to allow the user analysing
their eQTL data. This customisation (if provided) will populate
built-in eQTL data}

\item{include.RGB}{genes linked to input SNPs are also included. By
default, it is 'NA' to disable this option. Otherwise, those genes
linked to SNPs will be included according to Promoter Capture HiC
(PCHiC) datasets. Pre-built HiC datasets are detailed in
\code{\link{oDefineRGB}}}

\item{cdf.function}{a character specifying a Cumulative Distribution
Function (cdf). It can be one of 'exponential' based on exponential
cdf, 'empirical' for empirical cdf}

\item{relative.importance}{a vector specifying the relative importance
of nearby genes, eQTL genes and HiC genes. By default, it sets c(1/3,
1/3, 1/3)}

\item{scoring.scheme}{the method used to calculate seed gene scores
under a set of SNPs. It can be one of "sum" for adding up, "max" for
the maximum, and "sequential" for the sequential weighting. The
sequential weighting is done via: \eqn{\sum_{i=1}{\frac{R_{i}}{i}}},
where \eqn{R_{i}} is the \eqn{i^{th}} rank (in a descreasing order)}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{placeholder}{the characters to tell the location of built-in
RData files. See \code{\link{oRDS}} for details}

\item{guid}{a valid (5-character) Global Unique IDentifier for an OSF
project. See \code{\link{oRDS}} for details}
}
\value{
an object of class "xGene", a list with following components:
\itemize{
\item{\code{xGene}: a matrix with 3 columns; they are "Gene" (gene
symbols), "Score" (score values), and "Pval" (1e-10,1)}
\item{\code{SNP}: a data frame of nSNP X 4 containing input SNPs and/or
LD SNPs info, where nSNP is the number of input SNPs and/or LD SNPs,
and the 4 columns are "SNP" (dbSNP), "Score" (the SNP score), "Pval"
(the SNP p-value), "Flag" (indicative of Lead SNPs or LD SNPs)}
\item{\code{Gene2SNP}: a data frame of nPair X 3 containing Gene-SNP
pair info, where nPair is the number of Gene-SNP pairs, and the 3
columns are "Gene" (seed genes), "SNP" (dbSNP), "Score" (an SNP's
genetic influential score on a seed gene)}
\item{\code{nGenes}: if not NULL, it is a data frame containing
nGene-SNP pair info}
\item{\code{eGenes}: if not NULL, it is a data frame containing
eGene-SNP pair info per context}
\item{\code{cGenes}: if not NULL, it is a data frame containing
cGene-SNP pair info per context}
}
}
\description{
\code{oSNP2xGenes} is supposed to prioritise genes given a list of seed
SNPs together with the significance level. To prioritise genes, it
first defines and scores seed genes: nearby genes, eQTL genes and Hi-C
genes. With seed genes and their scores, it then uses Random Walk with
Restart (RWR) to calculate the affinity score of all nodes in the input
graph to the seed genes. The priority score is the affinity score.
Parallel computing is also supported for Linux-like or Windows
operating systems. It returns an object of class "pNode".
}
\note{
The scoring procedure (from SNPs to target genes) consists of following
steps:
\itemize{
\item{i) \code{\link{oSNPscores}} used to calculate the SNP score.}
\item{ii) \code{\link{oSNP2nGenes}} used to define and score the nearby
genes.}
\item{iii) \code{\link{oSNP2eGenes}} used to define and score the eQTL
genes.}
\item{iv) \code{\link{oSNP2cGenes}} used to define and score the HiC
genes.}
\item{v) define seed genes as the nearby genes in ii) and the eQTL
genes in iii) and the HiC genes in iv), which are then scored in an
integrative manner.}
\item{vi) \code{\link{oPierGenes}} used to prioritise genes using an
input graph and a list of seed genes and their scores from v). The
priority score is the affinity score estimated by Random Walk with
Restart (RWR), measured as the affinity of all nodes in the graph to
the seeds.}
}
}
\examples{

\dontrun{
# a) provide the SNPs with the significance info
ImmunoBase <- oRDS('ImmunoBase', placeholder=placeholder)
gr <- ImmunoBase$AS$variants
AS <- as.data.frame(gr) \%>\% select(Variant, Pvalue)

# b) perform priority analysis
pNode <- oSNP2xGenes(data=AS, include.QTL="eQTL_xMEdb_Monocyte",
include.RGB='PCHiC_PMID27863249_Combined', network="PCommonsUN_medium",
restart=0.7, placeholder=placeholder)

# c) save to the file called 'SNPs_priority.txt'
pNode$priority \%>\% write.delim("SNPs_priority.txt", delim="\t")

# d) manhattan plot
mp <- oPierManhattan(pNode, top=20, top.label.size=1.5, y.scale="sqrt",
placeholder=placeholder)
#pdf(file="Gene_manhattan.pdf", height=6, width=12, compress=TRUE)
print(mp)
#dev.off()
}
}
\seealso{
\code{\link{oSNPscores}}, \code{\link{oSNP2nGenes}},
\code{\link{oSNP2eGenes}}, \code{\link{oSNP2cGenes}},
\code{\link{oSparseMatrix}}, \code{\link{oSM2DF}},
\code{\link{oPierGenes}}, \code{\link{oSM2DF}},
\code{\link{oDefineQTL}}, \code{\link{oDefineRGB}}
}
